% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESS_t.R
\name{ESS_t}
\alias{ESS_t}
\title{Effective Sample Size for Tailoring}
\usage{
ESS_t(lambda, pi_u, t, distance_measure, epsilon, h_function = NULL, ...)
}
\arguments{
\item{lambda}{The discounting factor. A scalar greater or equal to zero. see 'Details'.}

\item{pi_u}{The unweighted probability of the outcome. A vector of equal size as the datapoints.
It must range between zero and one.}

\item{t}{The target threshold. A scalar between zero and one.}

\item{distance_measure}{The function to calculate the distance between pi_u and t. Possible values "squared"
or "absolute". Default is "squared", see 'Details.}

\item{epsilon}{The insensitivity threshold. A scalar greater or equal to zero. Default is zero, see 'Details'.}

\item{h_function}{A function object. Custom weighting function. Default is NULL.
Details section and the examples below for more information.}

\item{...}{Optional args.}
}
\value{
Scalar between 0 and 1.
}
\description{
Calculate the effective sample size for tailoring.
}
\details{
The Effective Sample Size is calculated as follows:

 \deqn{ESS_t = \sum_{i=1}{^n} w_i,}

 where \eqn{w_i = \exp - lambda h(pi_u, t)}

 and

 \eqn{i = 1, \ldots, n} (the sample size).

 The default function h() is the squared error with epsilon = 0, i.e

 \eqn{h(pi_u, t) = (pi_u - t)^2}.
}
