% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrop.R
\name{metrop_tailor}
\alias{metrop_tailor}
\title{Metropolis Algorithm}
\usage{
metrop_tailor(
  formula,
  data,
  lambda,
  pi_u,
  t,
  distance_measure,
  epsilon,
  h_function = NULL,
  initial,
  warm_up = 1000,
  burn_in = 1000,
  n_mcmc = 5000,
  tune = 2.38,
  iter_adapt = 200,
  user_prior_density = NULL,
  prior_mean = 0,
  prior_sd = 1000,
  proposal_function = proposal_fun,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula"(use link). The details of model specification are given under ‘Details’.}

\item{data}{Data frame.}

\item{lambda}{The discounting factor. A scalar greater or equal to zero. see 'Details'.}

\item{pi_u}{The unweighted probability of the outcome. A vector of equal size as the datapoints.
It must range between zero and one.}

\item{t}{The target threshold. A scalar between zero and one.}

\item{distance_measure}{The function to calculate the distance between pi_u and t. Possible values "squared"
or "absolute". Default is "squared", see 'Details.}

\item{epsilon}{The insensitivity threshold. A scalar greater or equal to zero. Default is zero, see 'Details'.}

\item{h_function}{A function object. Custom weighting function. Default is NULL.
Details section and the examples below for more information.}

\item{initial}{A real vector, the initial state of the Markov chain. If left unspecified a vector of zeros is used.}

\item{warm_up}{The number of warm-up iterations for the sampler.}

\item{burn_in}{The number of burn-in iterations for the sampler.}

\item{n_mcmc}{The number of Metropolis iterations for the sampler.}

\item{tune}{Scalar to achieve optimal acceptance rate. Default is 2.38;
\insertCite{@see Chapter 4, @brooks2011handbook}{TailorBayes}. Make sure that the acceptance rate is satisfactory
(typically between 0.20 and 0.5) before using the posterior sample for inference.}

\item{iter_adapt}{The number of iterations between warm-up and burn-in that the proposal is tuned. Default is 200.}

\item{user_prior_density}{If non-NULL, the prior (log)density up to a constant of proportionality.
This must be a function defined in R. See 'Details'.}

\item{prior_mean}{The mean for the default Gaussian prior. Default is 0.}

\item{prior_sd}{The standard deviation for the default Gaussian prior. Default is 1000.}

\item{proposal_function}{The proposal function. The default is \eqn{k}-variate Gaussian,
where \eqn{k} is the length of the parameter vector.}

\item{...}{Optional args.}
}
\value{
A list object where the first element \code{chain} is of class "mcmc". This element can be summarized by
functions provided by the coda package.
}
\description{
Metropolis Algorithm
}
\details{
\subsection{The Model}
  A typical predictor has the form response ~ terms where response is the binary
  response vector, denoted \eqn{y}, and terms is a series of terms (denoted \eqn{x})
  which specifies a linear predictor for response.

  \code{metrop_tailor()} simulates from the posterior distribution of a tailored logistic regression model
  using a random walk Metropolis algorithm. The model takes the following form:

  \deqn{y_i \sim \mathcal{B}ernoulli(\pi_i)}

   Where the inverse link function:

  \deqn{\pi_i = (\frac{\exp(x_i'\beta)}{1 + \exp(x_i'\beta)})^{w_i}}

  where \eqn{w_i = \exp - \{ \lambda h(pi_u, t, \epsilon) \} } and \eqn{i = 1, \ldots, n} (the sample size).
  The function \eqn{h()} is family  of \eqn{\epsilon}-insensitive functions, allowing customization using
  the \code{distance_measure} and \code{epsilon} arguments. The default function \eqn{h()} is
  the squared error with \eqn{\epsilon = 0}, i.e

  \eqn{h(pi_u, t, \epsilon) := (pi_u - t)^2}.

  The user can specify custom \eqn{h()}, see below. By default, we assume a multivariate Normal prior
  on \eqn{\beta}:

  \deqn{\beta \sim \mathcal{N}(0, I\sigma)}{\beta \sim \mathcal{N}(0, I\sigma),}

  with \eqn{\sigma = 1000}.

\subsection{Custom \code{h_function} and \code{user_prior_density}}
 Both use the function factory pattern \insertCite{@see Chapter 10, @wickham_advanced}{TailorBayes}.
 This is to avoid using either global variables or \dots arguments.
 See Examples below and \href{https://cran.r-project.org/web/packages/mcmc/vignettes/demo.pdf}{Geyer, 2020 Appendix A}.
}
\references{
\insertAllCited{}
}
