% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out}
\alias{out}
\title{Simulated logistic regression data.}
\format{
A data frame with 100 observations and 2 variables.

\describe{
 \item{x1}{quantitative predictor.}
 \item{x2}{quantitative predictor.}
 \item{y}{Bernoulli response.}
}
}
\usage{
out
}
\description{
As the title suggests.
}
\examples{
library(TailorBayes)
data(out)
fit <- glm(y ~ x1 + x2, family = binomial, data = out)
summary(fit)


}
\keyword{datasets}
