% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TailorBayes.R
\docType{package}
\name{TailorBayes}
\alias{TailorBayes}
\title{TailorBayes: A package for targeted Bayesian infrerence for binary outcomes}
\description{
The package provides functions for tailored Bayesian inference when different classification errors incur
different penalties. See vignette for an introduction.
}
\section{TailorBayes functions}{

The main function is \code{metrop_tailor()} which uses random walk MCMC algorithm to sample from the posterior.
It returns a list object where the first element, \code{chain}, is of class "mcmc". This element can be summarized by
functions provided by the coda package or similar ones. See function doc for details.

The \code{ESS_t()} functions calculates the effective sample size for tailoring, i.e., the sample size
inferences will be based on.
}

